% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIDOverlapMatrix.R
\name{calculateIDOverlapMatrix}
\alias{calculateIDOverlapMatrix}
\title{calculateIDOverlapMatrix}
\usage{
calculateIDOverlapMatrix(proteomicsMatrix)
}
\arguments{
\item{proteomicsMatrix}{Quantitative proteomics matrix with analyte ID in row.names and measurement id as column names.
Either NAs or 0s for missing values}
}
\value{
A distance matrix with ID set overlap in percent
}
\description{
Calculates % overlap (intersect/union) of ID sets from a proteomics matrix with identifiers in the row.names
}
