% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDiannResultSets.R
\name{compareDiannResultSets}
\alias{compareDiannResultSets}
\title{compareDiannResultSets}
\usage{
compareDiannResultSets(
  result_folder = "./",
  Venn_diagram = TRUE,
  compare_precursor_level = FALSE,
  heatmap = TRUE,
  sample_correlation = TRUE
)
}
\arguments{
\item{result_folder}{Path to the result folder. All resultsets identified by their *.stats.tsv file will be analyzed.}

\item{Venn_diagram}{Whether to calculate and draw Venn diagrams from the sets of analytes detected}

\item{compare_precursor_level}{Whether to include precursor level comparisons in Venn and heatmap. This might
be slow for calculation of heatmap, defaults to FALSE.}

\item{heatmap}{Whether to calculate and draw heatmaps, defaults to true. These are the most informative plots..}

\item{sample_correlation}{Whether to calculate sample-sample correlation and output heatmap and xy scatters on subset}
}
\value{
Returns a combined data.table of merged quantitative results with extra column resultset
}
\description{
Comparison of DIANN resultsets stored side-by-side in a result folder,
e.g. Results from pass1 analysis and pass2 re-analysis (Match Between Runs)
or results obtained using different input spectral libraries. Compares
the ID numbers from stats.tsv, but also the ID numbers obtained on protein group level
by counting from the quantitative tables. Compares sets on identified proteins in Venn Diagrams
and their quantitative patterns in side-by-side heatmap. Further, pairwise sample-sample correlation
for a subset of runs is implemented, however currently not experimental-design-aware.
}
