% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSets.R
\name{compareSets}
\alias{compareSets}
\title{compareSets}
\usage{
compareSets(
  sets_list = list(Set1 = c("A", "B", "C", "D", "E"), Set2 = c("A", "B", "D", "E",
    "F"), Set3 = c("B", "D", "E", "F", "G", "H")),
  venn_quantities = TRUE,
  plot_pdf = TRUE,
  pdf_analysis_tag = "current_analysis",
  plot_title = "set comparison"
)
}
\arguments{
\item{sets_list}{List of Character vectors with the respective sets. List names must be unique!}

\item{venn_quantities}{whether quantities (overlap size numbers) are plotted in Venn Diagrams. This is
recommended because the eulerr Venn Diagrams overlap areas are only with error proportional to actual overlap size.}

\item{plot_pdf}{selection whether pdf plots are created or whether the output occurs exclusively within Rconsole.}

\item{plot_title}{character selection either "R" or "pdf"}
}
\value{
returns a list of set comparison objects and plots from both UpSetR and Eulerr
}
\description{
Comparison of sets, e.g. peptides or proteins detected by a given analysis.
Draws both UpSetR and roughly area-proportional Venn Diagrams via the eulerr package.
For the comparison of more sets, see compareSets function.
}
