% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMzID2tsv.R
\name{convertMzID2tsv}
\alias{convertMzID2tsv}
\title{convertMzID2tsv}
\usage{
convertMzID2tsv(mzidfile)
}
\arguments{
\item{mzidfile}{path to .mzid file to be converted}
}
\value{
Writes mzidfile.tsv next to input file and returns a list of mzID R object and the corresponding flattened data.frame
}
\description{
Convert mzIdentML files to human-readable .tsv via the R/mzID package.
}
