% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getSetAnalysisGroups}
\alias{getSetAnalysisGroups}
\title{getSetAnalysisGroups}
\usage{
getSetAnalysisGroups(
  listInput = list(Set1 = c("A", "B", "C", "D", "E"), Set2 = c("A", "B", "D", "E", "F"),
    Set3 = c("B", "D", "E", "F", "G", "H")),
  sort = TRUE
)
}
\arguments{
\item{input}{list of named character vectors containing the sets}
}
\description{
Tracks members in intersect sets. Grabbed & modified from https://github.com/hms-dbmi/UpSetR/issues/85.
}
