% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDOverlapHeatmap.R
\name{IDOverlapHeatmap}
\alias{IDOverlapHeatmap}
\title{IDOverlapHeatmap}
\usage{
IDOverlapHeatmap(
  proteomicsMatrix,
  save_pdf = TRUE,
  pdf_suffix = "objname",
  size = 7
)
}
\arguments{
\item{proteomicsMatrix}{Quantitative proteomics matrix with analyte ID in row.names and measurement id as column names.
Either NAs or 0s for missing values}
}
\value{
A distance matrix with ID set overlap in percent
}
\description{
Calculates and visualizes in heatmap the % overlap (intersect/union) of ID sets from a proteomics matrix
}
