% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlynx2xvis.R
\name{maxlynx2xvis}
\alias{maxlynx2xvis}
\title{maxlynx2xvis
simplify maxlynx crosslinks table for visualization of inter- and intra-protein crosslinks via xvis
(https://xvis.genzentrum.lmu.de/CrossVisNoLogin.php)}
\usage{
maxlynx2xvis(
  combined_folder_location = "path/to/combined/",
  expDesign = NULL,
  xvis_filter = "InterProXL|IntraProXL",
  xvis_remove_decoys = TRUE,
  plot_pdf = TRUE
)
}
\arguments{
\item{combined_folder_location}{path to combined folder location. Character, either of length 1; light crosslinker only is assumed;
if two folders are provided (length 2) it is assumed that an isotopically labelled XL reagent was used; with two separate searches performed
for each channel. The first folder is assumed to be from the light search and the second folder is assumed to be from the heavy search.}

\item{expDesign}{In case there is no filled-in experimentalDesignTemplate.txt in the
combined folder, here, a data.table with column names 'Name' and 'Experiment'
should minimally be provided. Additional columns will be propagated to the output xvis_xls.csv}

\item{xvis_filter}{Filter string which Crosslink types should be included in the output xvis_xls.csv; defaults to "InterProXL|IntraProXL" (both included)}

\item{xvis_remove_decoys}{Logical, whether t-d, d-t and d-d XLs shall be removed before export to xvis_xls.csv. Defaults to TRUE.}

\item{plot_pdf}{Logical, whether overview pdf should be generated, highlighting the ID numbers of and overlap between channels}
}
\value{
a list of 1) merged data.table of crosslinks underlying xvis_xls.csv; 2) Crosslink MSMS count table underlying summary plot;
3) Summary ggplot
}
\description{
maxlynx2xvis
simplify maxlynx crosslinks table for visualization of inter- and intra-protein crosslinks via xvis
(https://xvis.genzentrum.lmu.de/CrossVisNoLogin.php)
}
