% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTicProfiles.R
\name{plotTicProfiles}
\alias{plotTicProfiles}
\title{plotTicProfiles}
\usage{
plotTicProfiles(
  path_to_raw = "../data/dda/",
  study_design = NULL,
  write_csv = "none"
)
}
\arguments{
\item{path_to_raw}{Path folder containing the .raw files}

\item{study_design}{Optional, if plots by condition are desired.
Study design table as data.table or path to tab-separated .txt with mandatory columns:
\itemize{
\item filename: raw file name including .raw extension
\item condition: String, biological condition (e.g. "treated" and "untreated")
\item replicate: Replicate number (integer).
}}

\item{write_csv}{Character, whether .csv output shall be written
\itemize{
\item next to raw files: "with_rawfiles"
\item in the current working directory: "workdir"
\item not at all: "none"
}}
}
\value{
list with elements tic_table (data.table) and plot (ggplot)
}
\description{
A function to extract and plot TIC profiles from Thermo.raw mass spectrometry data files
}
