% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFragPipeProteinTable.R
\name{splitFragPipeProteinTable}
\alias{splitFragPipeProteinTable}
\title{splitFragPipeProteinTable
Splits FragPipe combined_protein.tsv table into one table per quantification type}
\usage{
splitFragPipeProteinTable(
  combined_protein_tsv = "combined_protein.tsv",
  write_split_tables = TRUE
)
}
\arguments{
\item{write_split_tables}{Logical, whether split tables should be written, or only returned as elements of a list in R. Default: TRUE}

\item{combined_protein_file}{path to combined folder location. Character, either of length 1; light crosslinker only is assumed;
if two folders are provided (length 2) it is assumed that an isotopically labelled XL reagent was used; with two separate searches performed
for each channel. The first folder is assumed to be from the light search and the second folder is assumed to be from the heavy search.}
}
\value{
a list of the separate tables
}
\description{
splitFragPipeProteinTable
Splits FragPipe combined_protein.tsv table into one table per quantification type
}
